refactor(revision-diff-api): move orchestration from routes into services
Addresses kiro-review Finding 1 & 2 (PR #11353): the design mandates a thin
route layer that delegates to a service that owns the discovery/diff logic, but
the MongoDB orchestration (listChanges / computeDiffs) had been implemented
inline in the route files, leaving the declared ChangesIndexService /
RevisionDiffService interfaces as dead code.

- Move listChanges (cross-author aggregation, $setWindowFields, run grouping,
  keyset pagination, Page bulk access check) into changes-index-service.ts as
  the concrete ChangesIndexService.listChanges implementation.
- Move computeDiffs (per-batch bulk fetch + per-pair authorization) into
  revision-diff-service.ts as the concrete RevisionDiffService.computeDiffs
  implementation.
- Routes are now thin adapters: validate/normalize -> service -> res.apiv3.
- Remove the now-redundant empty ChangesIndexService / RevisionDiffService
  interfaces; keep NormalizedChangesQuery / NormalizedDiffRequest as the service
  param types and wire RevisionDiffRequest as the diff route body type.
- Drop the unused `crowi` param and `userId` locals; the service derives userId
  from the authenticated `user` (required by Page.findByIdsAndViewer).

Behavior-preserving: all 66 tests (38 unit + 28 integ) stay green; tsgo --noEmit
and biome pass.
